update benches to work on latest branch
