+++ .gitignore
@@ -9,6 +9,7 @@ Cargo.lock
 # These are backup files generated by rustfmt
 **/*.rs.bk
 
+*.cmx
 *.cmi
 *.cmo
 *.o
