+++ src/opt.rs
@@ -522,6 +522,7 @@ impl JProgram {
       flag_builder.set("use_colocated_libcalls", "false").unwrap();
       flag_builder.set("is_pic", "true").unwrap();
       flag_builder.set("enable_llvm_abi_extensions", "true").unwrap();
+      flag_builder.set("opt_level", "speed").unwrap();
       let isa_builder = cranelift_native::builder().unwrap_or_else(|msg| {
           panic!("host machine is not supported: {}", msg);
       });
